requirements.txt
django==5.0.7
django-filter==21.1
django-extensions==3.2.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
graphene-django==3.2.2
graphene-file-upload==1.3.0
graphene_django_filter==0.6.5
openai==1.39.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
psycopg2-binary==2.9.9
python-dotenv==1.0.1
mutagen==1.47.0
numpy==1.26.4
bpython==0.24.0
Redis==5.0.8

